+++ projects/quang-chart/src/lib/quang-chart.module.ts
@@ -11,6 +11,9 @@ import { ChartPieComponent } from './chart-pie/chart-pie.component'
 import { ChartRadarComponent } from './chart-radar/chart-radar.component'
 import { ChartTreemapComponent } from './chart-treemap/chart-treemap.component'
 import { TranslocoModule } from '@ngneat/transloco'
+import { ChartGauge } from './chart-gauge/chart-gauge.model'
+import { ChartGaugeComponent } from './chart-gauge/chart-gauge.component'
+import { ChartTreeComponent } from './chart-tree/chart-tree.component'
 
 @NgModule({
   declarations: [
@@ -21,7 +24,9 @@ import { TranslocoModule } from '@ngneat/transloco'
     ChartLineComponent,
     ChartPieComponent,
     ChartRadarComponent,
-    ChartTreemapComponent
+    ChartTreemapComponent,
+    ChartGaugeComponent,
+    ChartTreeComponent
   ],
   imports: [
     NgxEchartsModule.forRoot({ echarts }),
@@ -36,7 +41,9 @@ import { TranslocoModule } from '@ngneat/transloco'
     ChartLineComponent,
     ChartPieComponent,
     ChartRadarComponent,
-    ChartTreemapComponent
+    ChartTreemapComponent,
+    ChartGaugeComponent,
+    ChartTreeComponent
   ]
 })
 export class QuangChartModule { }