+++ projects/quang-components/src/lib/quang-components.module.ts
@@ -5,6 +5,8 @@ import { QuixCardSimpleComponent } from './quix-card-simple/quix-card-simple.com
 import { CommonModule } from '@angular/common'
 import { QuixCardHeaderComponent } from './quix-card-header/quix-card-header.component'
 import { QuixCardActionHeaderComponent } from './quix-card-action-header/quix-card-action-header.component'
+import { QuixCardImageComponent } from './quix-card-image/quix-card-image.component'
+import { TranslocoModule } from '@ngneat/transloco'
 
 @NgModule({
   declarations: [
@@ -12,17 +14,20 @@ import { QuixCardActionHeaderComponent } from './quix-card-action-header/quix-ca
     QuixCardActionComponent,
     QuixCardActionHeaderComponent,
     QuixCardSimpleComponent,
-    QuixCardHeaderComponent
+    QuixCardHeaderComponent,
+    QuixCardImageComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    TranslocoModule
   ],
   exports: [
     QuixCardComponent,
     QuixCardActionComponent,
     QuixCardActionHeaderComponent,
     QuixCardSimpleComponent,
-    QuixCardHeaderComponent
+    QuixCardHeaderComponent,
+    QuixCardImageComponent
   ]
 })
 export class QuangComponentsModule {}