+++ projects/quang-components/src/lib/quix-card-image/quix-card-image.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { QuixCardImageComponent } from './quix-card-image.component';
+
+describe('QuixCardImageComponent', () => {
+  let component: QuixCardImageComponent;
+  let fixture: ComponentFixture<QuixCardImageComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ QuixCardImageComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(QuixCardImageComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});