+++ projects/quang-components/src/public-api.ts
@@ -4,6 +4,9 @@
 
 export * from './lib/quix-card/quix-card.component'
 export * from './lib/quix-card-action/quix-card-action.component'
+export * from './lib/quix-card-action-header/quix-card-action-header.component'
+export * from './lib/quix-card-header/quix-card-header.component'
+export * from './lib/quix-card-image/quix-card-image.component'
 export * from './lib/quix-card-simple/quix-card-simple.component'
 
 export * from './lib/quang-components.module'