+++ CHANGELOG.md
@@ -16,3 +16,5 @@
 * update to eslint
 * update storage service for boolean management
 * change all selector from quix- to quang-
+
+## 12.